Update README.md
This project involved building a Decision Tree Classifier to predict income level (>50K or <=50K) based on various demographic and employment-related features from the dataset.

After loading and exploring the data, categorical features were encoded using Ordinal Encoding. The dataset was then split into training and testing sets. A Decision Tree Classifier was trained on the training data and evaluated on the test data.

Plotted 8 different type of visualization based on the income dataset.

The model achieved an accuracy of 89% on the test set. The confusion matrix and classification report show that the model performs well in classifying both income levels, with good precision, recall, and F1-scores for both the 'neutral' (<=50K) and 'positive' (>50K) classes. This indicates the model is reasonably effective in distinguishing between individuals with income above and below 50K based on the provided features.